Implement User authentication and authorization -User can only read, update, and delete their own account -services.py: will contain the business logic, currently only has get_user_by_email

+++ backend/app/users/router.py
@@ -14,8 +14,9 @@ from sqlalchemy.ext.asyncio import AsyncSession
 from app.auth.services import CurrentUser, get_password_hash
 from app.database.models import User as UserModel
 from app.database.sqlconnector import get_db
-from app.users.schema import UserCreate, UserPrivate, UserPublic, UserUpdate
-from app.utils import get_user_util
+from app.users.schema import UserCreate, UserPrivate, UserUpdate
+
+from . import services
 
 router = APIRouter()
 
@@ -36,14 +37,9 @@ async def register_user(
     """
 
     # Check if email has already been registered
-    result = await db.execute(
-        select(UserModel).where(
-            func.lower(UserModel.email_address) == user.email_address.lower()
-        ),
-    )
-    existing_user = result.scalar_one_or_none()
+    user = services.get_user(user.email_address, session=db)
 
-    if existing_user:
+    if user:
         raise HTTPException(
             status_code=status.HTTP_409_CONFLICT, detail="Email already registered."
         )
@@ -66,23 +62,12 @@ async def read_user_info_me(current_user: CurrentUser):
     return current_user
 
 
-@router.get("/{user_id}", response_model=UserPublic)
-async def get_user(user_id: UUID, db: Annotated[AsyncSession, Depends(get_db)]):
-    """Get a user by their id.
-
-    Args:
-        user_id (UUID): Id of user.
-        db (Annotated[AsyncSession, Depends): Session
-
-    Raises:
-        HTTPException: Raises 404 if user not found.
-    """
-    return await get_user_util(user_id, db)
-
-
 @router.patch("/{user_id}", response_model=UserPrivate)
 async def partial_update_user(
-    user_id: UUID, user_data: UserUpdate, db: Annotated[AsyncSession, Depends(get_db)]
+    user_id: UUID,
+    current_user: CurrentUser,
+    user_data: UserUpdate,
+    db: Annotated[AsyncSession, Depends(get_db)],
 ):
     """Update or partially update a user.
 
@@ -92,12 +77,20 @@ async def partial_update_user(
         db (Annotated[AsyncSession, Depends): session
 
     Raises:
-        HTTPException: Raises 404 if user not found.
         HTTPException: Raises 409 if email address is a duplicate.
+        HTTPException: Raises 404 if user not found.
     """
+    if user_id != current_user.id:
+        raise HTTPException(
+            status_code=status.HTTP_403_FORBIDDEN,
+            detail="Not authorized to delete this user.",
+        )
 
-    # Check of user exists
-    user = await get_user_util(user_id, db)
+    user = await db.get(UserModel, user_id)
+    if not user:
+        raise HTTPException(
+            status_code=status.HTTP_404_NOT_FOUND, detail="User not found."
+        )
 
     # only update fields that were sent in the request
     user_update_data = user_data.model_dump(exclude_unset=True)
@@ -130,7 +123,11 @@ async def partial_update_user(
 
 
 @router.delete("/{user_id}", status_code=status.HTTP_204_NO_CONTENT)
-async def delete_user(user_id: UUID, db: Annotated[AsyncSession, Depends(get_db)]):
+async def delete_user(
+    user_id: UUID,
+    current_user: CurrentUser,
+    db: Annotated[AsyncSession, Depends(get_db)],
+):
     """Delete a user.
 
     Args:
@@ -140,7 +137,18 @@ async def delete_user(user_id: UUID, db: Annotated[AsyncSession, Depends(get_db)
     Raises:
         HTTPException: Raises 404 if user not found.
     """
-    user = await get_user_util(user_id, db)
+
+    if user_id != current_user.id:
+        raise HTTPException(
+            status_code=status.HTTP_403_FORBIDDEN,
+            detail="Not authorized to delete this user.",
+        )
+
+    user = await db.get(UserModel, user_id)
+    if not user:
+        raise HTTPException(
+            status_code=status.HTTP_404_NOT_FOUND, detail="User not found."
+        )
 
     await db.delete(user)
     await db.commit()
